requirements.txt
pandas>=2.1
numpy>=1.26
pyarrow>=14.0
requests>=2.31
yfinance>=0.2
PyYAML>=6.0
scikit-learn>=1.4
xgboost>=2.0
hmmlearn>=0.3
pandas-market-calendars>=4.4
python-dotenv>=1.0
pytest>=8.0
tqdm>=4.66
statsmodels>=0.14
scipy>=1.12
matplotlib>=3.8
